rc_webservices/APIs/NewsService/requirements.txt
django~=2.2
djangorestframework==3.13.1
psycopg2==2.8.6
numpy==1.22.3